fix: add ability to upload images to the user.

+++ src/components/user/daos/user.mongoose.dao.ts
@@ -7,6 +7,8 @@ import { ICrudUser } from '../../../common/types/crud.interface';
 import { BadRequestError } from '../../../common/error/bad.request.error';
 import { User } from '../models/user.model';
 import { EPermissionLevel } from '../../../common/types/common.permissionlevel.enum';
+import MailService from '../../../services/mail/mail.service';
+import CloudinaryService from '../../../services/cloudinary/cloudinary.service';
 
 const log: debug.IDebugger = debug('app:users-dao');
 
@@ -21,13 +23,27 @@ class UsersDao implements ICrudUser {
         ...userFields,
         permissionLevel: EPermissionLevel.FREE_PERMISSION,
       });
+      // optional because the user can use the default image provided by the server
+      if (userFields.imageId && typeof userFields.imageId !== 'string') {
+        const { public_id, secure_url } = await CloudinaryService.uploadImage(
+          userFields.imageId,
+          'Products'
+        );
+        user.imageId = public_id;
+        user.imageUrl = secure_url;
+      }
       await user.save();
+      // we send an email to the administrator when a new user registers
+      MailService.send(
+        `New user ${userFields.firstName} - (${new Date().toLocaleString()})`,
+        userFields.toString()
+      );
       return user.id;
     } catch (err) {
       if (err instanceof mongoose.Error.ValidationError) {
         const message = Object.values(err.errors).map((prop) => prop.message);
         throw new BadRequestError(message.join('. '), 'create');
-      }
+      } else if (err instanceof BaseError) throw err;
       throw new BaseError('Failed to save user', err, 'create');
     }
   }
@@ -68,10 +84,17 @@ class UsersDao implements ICrudUser {
     }
   }
 
-  public async deleteById(userId: string) {
+  public async deleteById(user: ICreateUserDto) {
     try {
-      return User.deleteOne({ _id: userId }).exec();
+      if (
+        typeof user.imageId === 'string' &&
+        user.imageId !== 'Users/avatar.jpg'
+      ) {
+        await CloudinaryService.deleteImage(user.imageId);
+      }
+      return User.deleteOne({ _id: user.id }).exec();
     } catch (err) {
+      if (err instanceof BaseError) throw err;
       throw new BaseError('Failed to remove user', err, 'deleteById');
     }
   }
